Added the  "php-ldap" dependency
## Problem
- *Issue https://github.com/YunoHost-Apps/nextcloud_ynh/issues/69* 
```
Warning:   [Exception]
Warning:   App "LDAP user and group backend" cannot be installed because the following dependencies are not fulfilled: The library ldap is not available.
```
Dependency not installed : LDAP plugin of PHP.

## Solution
- *Install missing dependencies "php-ldap"*

## PR Status
To be reviewed. 

## Validation
---
*Minor decision*
- [ ] **Upgrade previous version** : 
- [ ] **Code review** : 
- [ ] **Approval (LGTM)** : 
- [ ] **Approval (LGTM)** : 
- [ ] **CI succeeded** : 
When the PR is mark as ready to merge, you have to wait for 3 days before really merge it.